removing useless components // refactors
