+++ client/src/components/shared/Navbar.tsx
@@ -2,7 +2,7 @@ import { Link } from 'react-router-dom';
 import { HiOutlineBars3BottomRight } from 'react-icons/hi2';
 import { IoGameController } from 'react-icons/io5';
 
-import Auth from '../utils/auth';
+import Auth from '../../utils/auth';
 
 const Navbar = () => {
 	return (
