--- client/src/pages/Home.tsx
@@ -1,12 +1,0 @@
-import Navbar from '../components/Navbar';
-import SearchGame from '../components/search/SearchGame';
-
-const Home = () => {
-	return (
-		<>
-			<Navbar />
-			<SearchGame />
-		</>
-	);
-};
-export default Home;
